Add uncategorized to categories
This one has the potential to be a performance hit; it should be
analyzed on a production-size database before deploying.  We may need to
figure out a good way to cache uncategorized tags along with regular --
perhaps by a special invalid tag that gets updated with a tag count by
the management command.