Add org status email
An email that can be fired by a cron job once every X days, which gives
you a snapshot of work that needs to be done.  Includes some playful
bits as well.

Also, remove auto_now from scans/documents to avoid the situation where
database imports don't preserve the older 'modified' time.  Use a save
method override instead.  Since we're using modified dates as part of
the representation of 'status', we want them to be accurate.