Redaction UI (no backend); switch coffee rendering
Filling in a UI for redactions in posts (backend not working yet).
While writing this, it became clear that the hackish autocompile script
for coffeescript wasn't good for development (compiler watcher would
crash periodically, leaving files not updated); so switch to compress
"precompilers" renderer.  Leaving compass on autocompile for now.