Denormalization bug in_reply_to; comment doc urls
The shadow comment created for a document when saving it with
in_reply_to was not updating its "parent" property, which was making the
comment point to the wrong parent if in_reply_to changed after the first
save.  Fix this.

Also, change the absolute url in "in_reply_to" documents so that it goes
to the comment rather than the top-level post view.